+++ ChangeLog.md
@@ -3,6 +3,7 @@
 # Version 0.2.0 (11/16/2014)
 
 - Add TravisCI and Drone.io for continuous integration
+- Implements Travis mode to display projects build status
 - Enable code coverage
 
 # Version 0.1.0 (11/14/2014)